requirements.txt
Flask
Flask-SQLAlchemy
SQLAlchemy
psycopg2-binary
python-dotenv
Flask-Login
gunicorn
python-escpos==3.1
Pillow>=11.0
pyusb
python-barcode
qrcode
